refactor(dashboard): remove private queries/auth.ts and wire dashboard actions to centralized @/lib/auth/roles guards
